chore(plugins): split plugin settings tests to own class
